requirements.txt
streamlit
Pillow
tensorflow
opencv-python-headless
numpy
